+++ CHANGELOG.md
@@ -8,6 +8,8 @@ e este projeto adere a [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 ### [Unreleased]
 
 ##### Added
+- Suporte ao Java 11.
+- Criação das classes **ThrowingRunnable**.
 - Criada anotação para testes na camada de serviço.
 - Implementação do recurso de _Multi Tenancy_.
 - Criação das classes **ThrowingConsumer** e **ThrowingFunction**.